training/_archived_old_approach/requirements.txt
# BrowerAI Model Training Requirements v2.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0
onnxscript>=0.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Modern NLP Tools (optional - for advanced tokenization)
# transformers>=4.30.0  # HuggingFace tokenizers
# tokenizers>=0.13.0    # Fast tokenizers

# Web Scraping (for data collection)
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# CSS/JS Processing
cssutils>=2.7.0
esprima>=4.0.0

# Data Augmentation
# nlpaug>=1.1.0  # Optional

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Model Export
onnxmltools>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest>=7.4.0
